Add files via upload
The "YouTube TIMESTAMPS" is a chrome extention project which aims to streamline the user experience of navigating and revisiting specific moments within YouTube videos by providing a simple and efficient bookmarking tool. The extension facilitates the saving and organization of timestamps within YouTube videos, allowing users to mark significant moments or points of interest for future reference.